Create functions that builds adjList Create findIndex that helps to find the start and end index
Create infoArr that adds objects containing distance and
predecessor to each index in the board

Create adjList that is made of neighbours for each
position in the board. Neighbours are the positions
that a knight can travel to.

Create a knightPossMoves which spawns all the possible
moves for a given position (maximum 8 per position).

This particular project was really too difficult. I
ended up looking up how other people completed the
project so that I could figure out how to do it.